% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloPlot.R
\name{run_elo_plot}
\alias{run_elo_plot}
\title{Run Elo plot}
\usage{
run_elo_plot(plot_teams = c("PHX", "OKC", "LAL", "BOS", "GSW", "MIA"))
}
\arguments{
\item{plot_teams}{Character vector of team abbreviations.}
}
\description{
Generates a season-reset Elo trajectory plot for selected teams.
}
